requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
idna==3.7
joblib==1.4.2
numpy==2.0.0
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
sqlparse==0.5.0
threadpoolctl==3.5.0
urllib3==2.2.2
